requirements.txt
pandas>=2.0
python-dateutil
requests
lzma